lmkd: allow devices to forcefully enable use_inkernel_interface
Some devices may want its kernel to handle lmk but has
ro.lmk.enable_userspace_lmk enabled in
/vendor/etc/perf/perfconfigstore.xml.

Allowing devices to opt-out in /vendor/build.prop is cleaner than
overriding /vendor/etc/perf/perfconfigstore.xml.

orges: this is not appliable to pure aosp because it is not reading the
prop in perfconfigstore.xml but instead it is reading minfree. anyways,
i'm forcing it to make sure

Change-Id: I5c83a9a07bbe03285512fc904498624960840e28
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>